removebg/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backgroundremover==0.2.6
beautifulsoup4==4.11.2
bitarray==2.8.0
bleach==6.0.0
blessed==1.20.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.0
cairocffi==1.5.0
certifi==2022.12.7
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1b0qzba5nr/croot/cffi_1670423213150/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
click-default-group==1.2.2
cloup==0.13.1
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.1.3
commandlines==0.4.1
conda==23.3.1
conda-content-trust @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0afea0f9-bbf5-446c-8d60-8151cc05b029p0nco1nn/croots/recipe/conda-content-trust_1658126375910/work
conda-package-handling @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_9an_ntrizz/croot/conda-package-handling_1666940386489/work
contourpy==1.0.7
controlnet-aux==0.0.5
coremltools==6.2
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_3evwafgyg8/croot/cryptography_1665612651044/work
cycler==0.11.0
Cython==0.29.34
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.14.0
easydict==1.10
EasyProcess==1.1
einops==0.6.1
entrypoint2==1.1
exceptiongroup==1.1.3
executing==1.2.0
fairseq==0.12.2
fastjsonschema==2.17.0
ffmpeg-python==0.2.0
filelock==3.10.0
filetype==1.2.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.0
future==0.18.3
gast==0.4.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
glcontext==2.3.7
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hsh==1.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.0.7
idna @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_00jf0h4zbt/croot/idna_1666125573348/work
imageio==2.30.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
inquirer==3.1.3
insightface==0.7.3
ipykernel==6.23.1
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isosurfaces==0.1.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
kiwisolver==1.4.4
labelImg==1.8.6
langchain==0.0.184
lazy_loader==0.2
libclang==16.0.0
litellm==0.1.610
llvmlite==0.41.1
lxml==4.9.2
manim==0.17.3
manim-ml==0.0.24
manimgl==1.6.1
ManimPango==0.4.3
mapbox-earcut==1.0.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.0.4
moderngl==5.8.2
moderngl-window==2.4.3
more-itertools==10.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multipledispatch==0.6.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.0.6
onnx==1.14.0
onnx-coreml==1.3
onnxruntime==1.15.0
open-interpreter==0.1.4
openai==0.27.10
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
pluggy @ file:///opt/concourse/worker/volumes/live/8277900c-164a-49c8-6f2a-f55c3c0154be/volume/pluggy_1648042581708/work
pooch==1.8.0
portalocker==2.7.0
prettytable==3.7.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycairo==1.23.0
pycosat @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ebsmghz9nu/croot/pycosat_1666805511853/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.8
pydub==0.25.1
PyGetWindow==0.0.9
pyglet==2.0.5
Pygments==2.14.0
PyGObject==3.44.1
PyMatting==1.1.12
PyMsgBox==1.0.9
pyobjc-core==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-Quartz==9.1.1
PyOpenGL==3.1.7
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
pyrr==0.10.3
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks @ file:///Users/builder/ci_310/pysocks_1642536366386/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
readchar==4.0.5
regex==2023.3.23
rembg==2.0.53
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_516b78ce-034d-4395-b9b5-1d78c2847384qtnol99l/croots/recipe/requests_1657734628886/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
ruamel.yaml @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_7de1zqcaou/croot/ruamel.yaml_1666304553877/work
ruamel.yaml.clib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c7s0zxy4t2/croot/ruamel.yaml.clib_1666302244557/work
rubicon-objc==0.4.6
sacrebleu==2.3.1
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
screeninfo==0.8.1
seaborn==0.12.2
selenium==4.11.2
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
skia-pathops==0.7.4
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.7
srt==3.5.3
stack-data==0.6.2
svgelements==1.9.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
timm==0.9.2
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
tokentrim==0.1.10
toolz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a7gkswah88/croot/toolz_1667464082910/work
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.2
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2adqcbsqqd/croots/recipe/tqdm_1664392689227/work
traitlets==5.9.0
transformers==4.27.4
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ffvqcdv424/croot/urllib3_1670526991804/work
usd-core==23.5
validators==0.22.0
waitress==2.1.2
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
